# internal/docs/swagger.yaml
definitions:
  common.JSONMap:
    additionalProperties: true
    type: object
...
  request.Empty:
    type: object
...
  request.GetAuthorityId:
    properties:
      authorityId:
        description: 角色ID
        type: integer
    type: object
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  request.GetUserList:
    properties:
      email:
        type: string
      keyword:
        description: 关键字
        type: string
      nickName:
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      phone:
        type: string
      username:
        type: string
    type: object
...
  request.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.SysMenusResponse:
    properties:
      menus:
        items:
          $ref: '#/definitions/system.SysMenu'
        type: array
    type: object
...
  system.Meta:
    properties:
      activeName:
        type: string
      closeTab:
        description: 自动关闭tab
        type: boolean
      defaultMenu:
        description: 是否是基础路由（开发中）
        type: boolean
      icon:
        description: 菜单图标
        type: string
      keepAlive:
        description: 是否缓存
        type: boolean
      title:
        description: 菜单名
        type: string
    type: object
...
  system.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      menus:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenu:
    properties:
      ID:
        description: 主键ID
        type: integer
      authoritys:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      component:
        description: 对应前端文件路径
        type: string
      createdAt:
        description: 创建时间
        type: string
      hidden:
        description: 是否在列表隐藏
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      meta:
        allOf:
          - $ref: '#/definitions/system.Meta'
        description: 附加属性
      name:
        description: 路由name
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由path
        type: string
      sort:
        description: 排序标记
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuBtn:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      desc:
        type: string
      name:
        type: string
      sysBaseMenuID:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuParameter:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      key:
        description: 地址栏携带参数的key
        type: string
      sysBaseMenuID:
        type: integer
      type:
        description: 地址栏携带参数为params还是query
        type: string
      updatedAt:
        description: 更新时间
        type: string
      value:
        description: 地址栏携带参数的值
        type: string
    type: object
...
  system.SysMenu:
    properties:
      ID:
        description: 主键ID
        type: integer
      authoritys:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      btns:
        additionalProperties:
          type: integer
        type: object
      children:
        items:
          $ref: '#/definitions/system.SysMenu'
        type: array
      component:
        description: 对应前端文件路径
        type: string
      createdAt:
        description: 创建时间
        type: string
      hidden:
        description: 是否在列表隐藏
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      menuId:
        type: integer
      meta:
        allOf:
          - $ref: '#/definitions/system.Meta'
        description: 附加属性
      name:
        description: 路由name
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由path
        type: string
      sort:
        description: 排序标记
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  system.SysUser:
    properties:
      ID:
        description: 主键ID
        type: integer
      authorities:
        description: 多用户角色
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      authority:
        allOf:
          - $ref: '#/definitions/system.SysAuthority'
        description: 用户角色
      authorityId:
        description: 用户角色ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      enable:
        description: 用户是否被冻结 1正常 2冻结
        type: integer
      headerImg:
        description: 用户头像
        type: string
      nickName:
        description: 用户昵称
        type: string
      originSetting:
        allOf:
          - $ref: '#/definitions/common.JSONMap'
        description: 配置
      phone:
        description: 用户手机号
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户登录名
        type: string
      uuid:
        description: 用户UUID
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.7.9-beta
paths:
...
  /menu/addBaseMenu:
    post:
      consumes:
        - application/json
      parameters:
        - description: 路由path, 父菜单ID, 路由name, 对应前端文件路径, 排序标记
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysBaseMenu'
      produces:
        - application/json
      responses:
        "200":
          description: 新增菜单
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 新增菜单
      tags:
        - Menu
...
  /menu/deleteBaseMenu:
    post:
      consumes:
        - application/json
      parameters:
        - description: 菜单id
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.GetById'
      produces:
        - application/json
      responses:
        "200":
          description: 删除菜单
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 删除菜单
      tags:
        - Menu
...
  /menu/getMenu:
    post:
      parameters:
        - description: 空
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.Empty'
      produces:
        - application/json
      responses:
        "200":
          description: 获取用户动态路由,返回包括系统菜单详情列表
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.SysMenusResponse'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
        - AuthorityMenu
  /menu/getMenuAuthority:
    post:
      consumes:
        - application/json
      parameters:
        - description: 角色ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.GetAuthorityId'
      produces:
        - application/json
      responses:
        "200":
          description: 获取指定角色menu
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    additionalProperties: true
                    type: object
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 获取指定角色menu
      tags:
        - AuthorityMenu
  /menu/getMenuList:
    post:
      consumes:
        - application/json
      parameters:
        - description: 页码, 每页大小
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.PageInfo'
      produces:
        - application/json
      responses:
        "200":
          description: 分页获取基础menu列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.PageResult'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 分页获取基础menu列表
      tags:
        - Menu
  /menu/updateBaseMenu:
    post:
      consumes:
        - application/json
      parameters:
        - description: 路由path, 父菜单ID, 路由name, 对应前端文件路径, 排序标记
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysBaseMenu'
      produces:
        - application/json
      responses:
        "200":
          description: 更新菜单
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 更新菜单
      tags:
        - Menu
...
  /user/SetSelfSetting:
    put:
      consumes:
        - application/json
      parameters:
        - description: 用户配置数据
          in: body
          name: data
          required: true
          schema:
            additionalProperties: true
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: 设置用户配置
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    additionalProperties: true
                    type: object
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 设置用户配置
      tags:
        - SysUser
...
  /user/deleteUser:
    delete:
      consumes:
        - application/json
      parameters:
        - description: 用户ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.GetById'
      produces:
        - application/json
      responses:
        "200":
          description: 删除用户
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 删除用户
      tags:
        - SysUser
...
  /user/getUserList:
    post:
      consumes:
        - application/json
      parameters:
        - description: 页码, 每页大小
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.GetUserList'
      produces:
        - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.PageResult'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 分页获取用户列表
      tags:
        - SysUser
  /user/resetPassword:
    post:
      parameters:
        - description: ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysUser'
      produces:
        - application/json
      responses:
        "200":
          description: 重置用户密码
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 重置用户密码
      tags:
        - SysUser
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...